requirements.txt
django
Faker
djangorestframework